requirements.txt
ultralytics
opencv-python
manga-ocr
numpy
translators
pillow
torch
torchvision
torchaudio